failed valid input and chaning mathod signature as excepted

+++ src/main/java/com/prasad/StringUtil.java
@@ -2,5 +2,5 @@ package com.prasad;
 
 public interface StringUtil {
 
-    public Double sumOfDigits(String str);
+    public int sumOfDigits(String str);
 }
